# Speed up reductions on non-contiguous dimensions