I just did `sum`, `prod`, `mean`, `max`, and `min` for now, but you can do a bunch of the other pointwise reductions like this too. 

before:
```
th> t = torch.randn(64,1024,1024)
                                                                      [5.6855s]
th> t:max(1);
                                                                      [1.7969s]
th> t:max(2);
                                                                      [0.9834s]
th> t:max(3);
                                                                      [0.1487s]
th> t:sum(1);
                                                                      [1.5895s]
th> t:sum(2);
                                                                      [0.9665s]
th> t:sum(3);
                                                                      [0.1610s]
```

after:
```
th> t = torch.randn(64,1024,1024)
                                                                      [5.7820s]
th> t:max(1);
                                                                      [0.3097s]
th> t:max(2);
                                                                      [0.2392s]
th> t:max(3);
                                                                      [0.1456s]
th> t:sum(1);
                                                                      [0.1985s]
th> t:sum(2);
                                                                      [0.2219s]
th> t:sum(3);
```

I think the optimizations for `max` and `min` will conflict with the TH_TENSOR_APPLY improvements in https://github.com/torch/torch7/pull/852, but we can always add special macros for this purpose if we want to keep the max and min optimizations. That new macro could also support stuff like `cumsum`.