@soumith 2017-02-10
it looks like we're lucky that TH_TENSOR_DIM_APPLY2 etc. dont have OpenMP optimizations yet. in the longer term, we need tree reductions.