@adamlerer 2017-02-10
@soumith re OpenMP: as I said to @apaszke this comes down to the specification of `TH_TENSOR_APPLY` (not TH_TENSOR_DIM_APPLY) w.r.t. stride-0 dimensions. Currently it handles them correctly. If someone changes it in a way that breaks that, they'll need to maintain a version that handles stride-0 correctly. Note that there's already existing code that sort of depends on handling this case correctly, e.g.

```
t = torch.Tensor(10, 10, 10)
u = torch.Tensor(1, 10, 10).expand_as(t).zero_()
u.add_(t)
```

so I'm not sure if it is now safe to break that (considering the specification was ill-defined).