@apaszke 2017-02-21
I think we could leave them as TH_TENSOR_APPLY_COUNTER or sth like that. They're all unacceptably slow, so let's keep the fix.