@adamlerer 2017-02-21
Alright, I rebased this on top of https://github.com/torch/torch7/pull/852, refactored TH_TENSOR_APPLY to remove all the copypasta and added a variant that can handle these reduction cases.

You'll have to wait until https://github.com/torch/torch7/pull/852 is landed to review this.