@soumith 2017-02-28
thanks Adam!